# docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GymCRM API",
    "description": "Multi-tenant gym SaaS. All business endpoints require Bearer JWT. gym_id is extracted from JWT — every query is automatically scoped to the authenticated gym.",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter: Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/v1/plans": {
      "post": { "tags": ["plans"], "summary": "Create a membership plan",
        "description": "Plan names are case-insensitively unique within a gym. Price in paise (₹1,500 = 150000).",
        "security": [{ "BearerAuth": [] }], "consumes": ["application/json"], "produces": ["application/json"],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/CreatePlanRequest" } }],
        "responses": {
          "201": { "description": "Created", "schema": { "$ref": "#/definitions/PlanResponse" } },
          "409": { "description": "Plan name already exists in this gym" },
          "422": { "description": "Validation error" },
          "401": { "description": "Not authenticated" }
        }
      },
      "get": { "tags": ["plans"], "summary": "List membership plans",
        "description": "Paginated list. Filter by active status or search by name.",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "page", "type": "integer" },
          { "in": "query", "name": "per_page", "type": "integer" },
          { "in": "query", "name": "search", "type": "string", "description": "Search by plan name" },
          { "in": "query", "name": "active", "type": "boolean", "description": "Filter by active status" }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/PlanListResponse" } } }
      }
    },
    "/api/v1/plans/active": {
      "get": { "tags": ["plans"], "summary": "List active plans",
        "description": "Returns only active, assignable plans. Use to populate plan dropdowns in Flutter. Sorted alphabetically.",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "page", "type": "integer" },
          { "in": "query", "name": "per_page", "type": "integer" }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/PlanListResponse" } } }
      }
    },
    "/api/v1/plans/{id}": {
      "get": { "tags": ["plans"], "summary": "Get plan by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "type": "integer", "required": true }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/PlanResponse" } }, "404": { "description": "Not found" } }
      },
      "put": { "tags": ["plans"], "summary": "Update a plan",
        "description": "Updates plan fields. Setting is_active=false hides the plan. Existing members and renewals are unaffected.",
        "security": [{ "BearerAuth": [] }], "consumes": ["application/json"],
        "parameters": [
          { "in": "path", "name": "id", "type": "integer", "required": true },
          { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/UpdatePlanRequest" } }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/PlanResponse" } }, "404": { "description": "Not found" }, "409": { "description": "Name taken" } }
      },
      "delete": { "tags": ["plans"], "summary": "Delete a plan (soft delete)",
        "description": "Soft-deletes the plan. Existing member and renewal records are preserved. Plan name becomes reusable.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "type": "integer", "required": true }],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    }
  },
  "definitions": {
...
    "CreatePlanRequest": {
      "type": "object", "required": ["name", "duration_days", "price_in_paise"],
      "properties": {
        "name": { "type": "string", "example": "Monthly Basic" },
        "description": { "type": "string", "example": "Access to all equipment" },
        "duration_days": { "type": "integer", "example": 30, "description": "Must be > 0" },
        "price_in_paise": { "type": "integer", "example": 150000, "description": "In paise. ₹1,500 = 150000" }
      }
    },
    "UpdatePlanRequest": {
      "type": "object",
      "properties": {
        "name": { "type": "string" }, "description": { "type": "string" },
        "duration_days": { "type": "integer" }, "price_in_paise": { "type": "integer" },
        "is_active": { "type": "boolean" }
      }
    },
    "PlanResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" }, "gym_id": { "type": "integer" },
        "name": { "type": "string" }, "description": { "type": "string" },
        "duration_days": { "type": "integer" },
        "price_in_paise": { "type": "integer", "example": 150000 },
        "price_in_rupees": { "type": "number", "example": 1500.0, "description": "Display only — use price_in_paise for calculations" },
        "is_active": { "type": "boolean" },
        "created_at": { "type": "string" }, "updated_at": { "type": "string" }
      }
    },
    "PlanListResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "data": { "type": "object", "properties": {
          "plans": { "type": "array", "items": { "$ref": "#/definitions/PlanResponse" } }
        }},
        "meta": { "type": "object", "properties": {
          "page": { "type": "integer" }, "per_page": { "type": "integer" },
          "total": { "type": "integer" }, "total_pages": { "type": "integer" }
        }}
      }
    }
  }
}